Create num_of_ways_with_obstacles_DP.py
There is an  grid in which there are  rows and  colums. A cell  is defined as the  row from the top and  column from left. You are located at  initially and can perform the following steps any number of times:

1. You can move *any* number of steps downwards.
2. You can move *any* number of steps towards the right.
There are some obstacles in the path.

You are initially located at  and wants to reach  but you are interested in knowing the number of ways you can reach  from . Since these numbers can huge, print it modulo .

Two ways are considered different if they have a different number of steps or differ in some positions.